Fix user preferences and refactor API endpoints
- Set default profile visibility to private for new users
- Refactor formats API endpoint with cleaner separation of concerns
- Fix user signup game preferences not being saved to database
- Update database trigger to handle both signup and login game preference formats
- Minor comment update in TrainingCourtWelcomeClient
