+++ app/api/formats/_utils.ts
@@ -1,5 +1,5 @@
-import { isAfter, isBefore } from "date-fns";
-import { PokemonSet } from "./_types";
+import { isAfter, isBefore, parseISO, nextFriday, add } from "date-fns";
+import { PokemonSet, PokemonTCGApiPokemonSet } from "./_types";
 
 const rotations = [
   new Date('2024-04-05'),
@@ -25,4 +25,34 @@ export const getRotationBlocks = (sets: PokemonSet[]) => {
       legalityDate: set.legalityDate,
     }
   }).filter((set) => set);
+}
+
+export const getProcessedPokemonSets = async (): Promise<PokemonSet[]> => {
+  const response = await fetch('https://api.pokemontcg.io/v2/sets');
+  const data: Record<string, PokemonTCGApiPokemonSet[]> = await response.json();
+  
+  const dataWithLegalities = data['data'].map((obj) => {
+    const releaseDate = parseISO(obj['releaseDate'].replaceAll('/', ''));
+    const legalityDate = add(nextFriday(releaseDate), {
+      weeks: 1
+    });
+
+    return ({
+      ...obj,
+      releaseDate: releaseDate.toISOString(),
+      legalityDate: legalityDate.toISOString()
+    })
+  })
+  
+  const alteredData = dataWithLegalities.sort((a, b) => {
+    if (isBefore(a.releaseDate, b.releaseDate)) return 1;
+    if (isAfter(a.releaseDate, b.releaseDate)) return -1;
+
+    if (a.name.length < b.name.length) return -1;
+    if (a.name.length > b.name.length) return 1;
+
+    return 0;
+  }).filter((set, idx) => dataWithLegalities.findIndex(({ legalityDate }) => legalityDate === set.legalityDate) === idx && !['sve'].includes(set.id) && !set.name.includes(' Promos') && set.series !== 'Other');
+
+  return alteredData;
 }
